+++ src/view/index.ts
@@ -6,5 +6,6 @@ export * from './web-component';
 export * from './media';
 export * from './menu-position';
 export * from './block-handles';
+export * from './drop-cursor';
 export * from './virtual-layout';
 export * from './paste';